REFACTOR : extract PostgresqlTestContainerBase to be an abstract class
