+++ src/test/java/com/expense_tracker/config/PostgresqlTestContainerBase.java
@@ -1,0 +1,37 @@
+package com.expense_tracker.config;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.test.context.ActiveProfiles;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+@Testcontainers
+@ActiveProfiles("test")
+public abstract class PostgresqlTestContainerBase {
+
+	@LocalServerPort
+	private Integer port;
+
+	@Container
+	@ServiceConnection
+	static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>("postgres:16-alpine");
+
+	@BeforeAll
+	static void beforeAll() {
+		// Call autmaticaly before any test to start postgre
+		postgres.start();
+	}
+
+	@AfterAll
+	static void afterAll() {
+		postgres.stop();
+	}
+
+}
